+++ changelog.md
@@ -8,6 +8,8 @@
 * refactor the debug pane in the fhirpath test page to use the standard resource editor/viewer
 * update the resource editor to use the <enter> key to download the resource, and also handle XML download better
 * add the Python R5 engine
+* split the fhirpath engine selector into 2 fields, one for the engine and one for the FHIR version
+* add interim version of the fhirpath.js running on R6
 
 # 1 August 2025 (dev)
 * Fix issue where json path walker was not correctly handling the `BackboneElement` and `Element` types in the json parser